Functionalize basic structure matching in silico
