+++ pydiffexp/pipeline.py
@@ -4,6 +4,7 @@ import multiprocessing as mp
 import os
 import shutil
 import tarfile
+import itertools as it
 
 import matplotlib.pyplot as plt
 import seaborn as sns
@@ -12,9 +13,11 @@ import numpy as np
 import pandas as pd
 from pydiffexp import DEAnalysis, DEResults, DEPlot, get_scores, cluster_discrete, pairwise_corr
 from pydiffexp.gnw import mk_ch_dir, GnwNetResults, GnwSimResults, draw_results, get_graph
+from pydiffexp.gnw.sim_explorer import tsv_to_dg, to_gephi
 from scipy import stats
 from scipy.spatial.distance import hamming
 from sklearn.metrics import mean_squared_error as mse
+from palettable.cartocolors import qualitative
 
 
 class DynamicDifferentialExpression(object):
@@ -88,7 +91,7 @@ class DynamicDifferentialExpression(object):
         return {'perturbation': 1, 'gene': 'y'}
 
     def score(self, prefix, exp, ctrl, sim_predictions=None, sim_filter=None,
-              f=None, reduced_set=True):
+              f=None, reduced_set=True, plot=False):
 
         if f is None:
             f = mse
@@ -135,7 +138,81 @@ class DynamicDifferentialExpression(object):
         g = match.groupby('true_gene')
 
         test_stats = pd.concat([self.dde_genes, g.apply(self.sample_stats, gene_mae_dist_dict, true_lfc, pred_lfc)], axis=1).dropna()
-        self.plot_results(test_stats)
+
+        if plot:
+            self.plot_results(test_stats)
+
+        return test_stats
+
+    @staticmethod
+    def load_structures(path):
+        return pd.read_csv(path)
+
+    @staticmethod
+    def expected_edges(node_dict, net: nx.DiGraph):
+        possible_edges = list(it.permutations(node_dict.values(), 2))
+        true_graph = nx.DiGraph()
+        for edge in possible_edges:
+            if nx.has_path(net, edge[0], edge[1]):
+                true_graph.add_edge(edge[0], edge[1])
+        return true_graph
+
+    def score_topology(self, true_topology):
+        grouped = self.estimators
+        ii = 0
+        max_edges = 6
+        summary = pd.DataFrame()
+        true_topology = '/Users/jfinkle/Documents/Northwestern/MoDyLS/Code/Python/pydiffexp/data/insilico/strongly_connected/Yeast-100_anonymized.tsv'
+        df, dg = tsv_to_dg(true_topology)
+        sim_path = '../data/motif_library/gnw_networks/'
+        edge_dict = {'x': 'G25', 'G': 'G38'}
+        for gene, data in grouped:
+            edge_dict['y'] = gene
+            tg = self.expected_edges(edge_dict, dg)
+            te = len(tg.edges())
+            ne = max_edges - te
+            gene_stats = []
+            for row, info in data.iterrows():
+                # print(info['id'])
+                _, net = tsv_to_dg("{path}/{id}/{id}_goldstandard_signed.tsv".format(path=sim_path, id=info.name[0]))
+                tp = []
+                fp = 0
+                net_edges = [(edge_dict[e[0]], edge_dict[e[1]]) for e in net.edges()]
+                for edge in net_edges:
+                    try:
+                        tp.append(nx.shortest_path_length(dg, edge[0], edge[1]))
+                    except:
+                        fp += 1
+                fn = 0
+                for edge in tg.edges():
+                    if edge not in net_edges:
+                        fn += 1
+                tn = max(0, ne - fn)
+                mean_true_path_length = np.mean(tp)
+                ntp = len(tp)
+                gene_stats.append([ntp / te, ntp + fp, fp / (ntp + fp), 2 * ntp / (2 * ntp + fp + fn),
+                                   (ntp + tn) / (ntp + fp + fn + tn), mean_true_path_length])
+            gene_data = pd.DataFrame(gene_stats, data.index.get_level_values('id').values,
+                                     columns=['TPR', 'total_edges', 'FDR', 'F1', 'ACC', 'mean_tp_length'])
+            gene_mean = pd.DataFrame(gene_data.mean())
+            gene_mean.columns = [gene]
+            gene_mean = gene_mean.T
+            gene_mean['true_edges'] = te
+            summary = pd.concat([summary, gene_mean], axis=0)
+            ii += 1
+        print(summary)
+        colors = qualitative.Bold_10.mpl_colors
+        plt.figure(figsize=(5, 8))
+        sns.set_style("whitegrid")
+        sns.boxplot(data=summary.loc[:, ['TPR', 'FDR', 'ACC', 'F1']], width=0.5, palette=colors, showfliers=False)
+        sns.swarmplot(data=summary.loc[:, ['TPR', 'FDR', 'ACC', 'F1']], color='0.2')
+        # plt.plot([-1,3], [0.5, 0.5], '--', c=colors[2], label='Random')
+        plt.ylim([0, 1])
+        # plt.xticks(rotation='vertical')
+        # plt.ylabel('Value')
+        plt.tight_layout()
+        plt.show()
+        sys.exit()
 
     @staticmethod
     def plot_results(x):
